% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{meili_index_create_filters}
\alias{meili_index_create_filters}
\title{Create filter(s) on Meilisearch index}
\usage{
meili_index_create_filters(index, fields)
}
\arguments{
\item{index}{the index to create filters in}

\item{fields}{the fields to create filters on}
}
\description{
Create filter(s) on Meilisearch index
}
