% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{search_name}
\alias{search_name}
\title{Search for a name in hrfile (or other index)}
\usage{
search_name(q, index = "hrfile")
}
\arguments{
\item{q}{the string to search for}

\item{index}{the index to search in, default hrfile}
}
\description{
Search for a name in hrfile (or other index)
}
