% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{simple_search}
\alias{simple_search}
\title{Simple search in Meilisearch index}
\usage{
simple_search(index, q, f = NULL)
}
\arguments{
\item{index}{the index to search in}

\item{q}{a string or list of strings to search for}

\item{f}{filter(s) to apply before search (optional)}
}
\description{
Simple search in Meilisearch index
}
