% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{meili_ingest_csv}
\alias{meili_ingest_csv}
\title{Ingest data into Meili index}
\usage{
meili_ingest_csv(index, csvfile, primary_key)
}
\arguments{
\item{index}{the name of the index to ingest data in}

\item{csvfile}{the name of a dataframe or a csv file to ingest into index}

\item{primary_key}{the field to use as primary key in the index}
}
\description{
Ingest data into Meili index
}
